Move data source for languges view into a separate file
removing from languages manager
