+++ godtools/Managers/LanguagesManager.swift
@@ -125,47 +125,3 @@ class LanguagesManager: GTDataManager {
                               .appendingPathComponent(self.path)
     }
 }
-
-extension LanguagesManager: LanguageTableViewCellDelegate {
-    func deleteButtonWasPressed(_ cell: LanguageTableViewCell) {
-        self.recordLanguageShouldDelete(language: cell.language!)
-    }
-    
-    func downloadButtonWasPressed(_ cell: LanguageTableViewCell) {
-        self.recordLanguageShouldDownload(language: cell.language!)
-        TranslationZipImporter.shared.download(language: cell.language!)
-    }
-}
-
-extension LanguagesManager: UITableViewDataSource {
-    
-    static let languageCellIdentifier = "languageCell"
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let language = languages[indexPath.row]
-        let cell = tableView.dequeueReusableCell(withIdentifier: LanguagesManager.languageCellIdentifier) as! LanguageTableViewCell
-        
-        cell.cellDelegate = self
-        cell.language = language
-        
-        return cell
-    }
-    
-    func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
-        let language = languages[indexPath.row]
-        let selected = language.remoteId == self.selectedLanguageId()
-        
-        if selected {
-            cell.setSelected(selected, animated: true)
-            tableView.selectRow(at: indexPath, animated: true, scrollPosition: .none)
-        }
-    }
-    
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return 1
-    }
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return languages.count
-    }
-}
