+++ godtools/ViewControllers/Platform/LanguagesTableViewController.swift
@@ -14,13 +14,14 @@ protocol LanguagesTableViewControllerDelegate {
 
 class LanguagesTableViewController: BaseViewController {
     
+    static let languageCellIdentifier = "languageCell"
+    
     var delegate: LanguagesTableViewControllerDelegate?
     
     var languages = List<Language>()
-    
     let languagesManager = LanguagesManager()
-    
     var screenTitleAux: String = "primary_language"
+    
     override var screenTitle: String {
         get {
             return screenTitleAux.localized
@@ -30,7 +31,7 @@ class LanguagesTableViewController: BaseViewController {
     @IBOutlet weak var tableView: UITableView! {
         didSet {
             tableView.delegate = self
-            tableView.dataSource = languagesManager
+            tableView.dataSource = self
         }
     }
     
@@ -58,8 +59,20 @@ class LanguagesTableViewController: BaseViewController {
             self.screenTitleAux = "parallel_language"
         }
     }
-    
+
     fileprivate func registerCells() {
-        self.tableView.register(UINib(nibName: "LanguageTableViewCell", bundle: nil), forCellReuseIdentifier: LanguagesManager.languageCellIdentifier)
+        self.tableView.register(UINib(nibName: "LanguageTableViewCell", bundle: nil),
+                                forCellReuseIdentifier: LanguagesTableViewController.languageCellIdentifier)
+    }
+}
+
+extension LanguagesTableViewController: LanguageTableViewCellDelegate {
+    func deleteButtonWasPressed(_ cell: LanguageTableViewCell) {
+        languagesManager.recordLanguageShouldDelete(language: cell.language!)
+    }
+    
+    func downloadButtonWasPressed(_ cell: LanguageTableViewCell) {
+        languagesManager.recordLanguageShouldDownload(language: cell.language!)
+        TranslationZipImporter.shared.download(language: cell.language!)
     }
 }
